AeroDocMVP/backend/rag/requirements.txt
qdrant-client>=1.9.0
docling
chonkie
fastembed
python-dotenv
tqdm
